move rag module to example
